+++ src/nl/hannahsten/texifyidea/run/latex/logtab/LatexOutputListener.kt
@@ -3,6 +3,10 @@ package nl.hannahsten.texifyidea.run.latex.logtab
 import com.intellij.execution.process.ProcessEvent
 import com.intellij.execution.process.ProcessListener
 import com.intellij.execution.process.ProcessOutputType
+import com.intellij.openapi.actionSystem.*
+import com.intellij.openapi.actionSystem.ex.ActionUtil
+import com.intellij.openapi.actionSystem.impl.SimpleDataContext
+import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.vfs.VirtualFile
@@ -15,6 +19,7 @@ import nl.hannahsten.texifyidea.run.latex.logtab.ui.LatexCompileMessageTreeView
 import nl.hannahsten.texifyidea.util.files.findFile
 import nl.hannahsten.texifyidea.util.remove
 import nl.hannahsten.texifyidea.util.removeAll
+import nl.hannahsten.texifyidea.util.runInBackgroundWithoutProgress
 import org.apache.commons.collections.Buffer
 import org.apache.commons.collections.BufferUtils
 import org.apache.commons.collections.buffer.CircularFifoBuffer
@@ -97,6 +102,33 @@ class LatexOutputListener(
 
             // Newlines are important to check when message end. Keep.
             processNewText(collectedLine)
+
+            checkForPdfRefresh(collectedLine)
+        }
+    }
+
+    /**
+     * latexmk -pvc remains running continously and compiles on file changes, so we need to refresh the pdf based on latexmk log messages since the pdf update
+     * will not appear in vfs until refreshed
+     */
+    fun checkForPdfRefresh(text: String) {
+        if (!"\\s*=== Watching for updated files. Use ctrl/C to stop ...\\s*".toRegex().matches(text)) {
+            return
+        }
+        runInBackgroundWithoutProgress {
+            val action = ActionManager.getInstance().getAction("pdf.viewer.ReloadViewAction") ?: return@runInBackgroundWithoutProgress
+            val event = AnActionEvent(
+                SimpleDataContext.getProjectContext(project),
+                Presentation(),
+                ActionPlaces.UNKNOWN,
+                ActionUiKind.NONE,
+                null,
+                0,
+                ActionManager.getInstance(),
+            )
+            runInEdt {
+                ActionUtil.performActionDumbAwareWithCallbacks(action, event)
+            }
         }
     }